[DS-1011] Handling enum vs numeric in IntegerizeExplodedSession

+++ src/main/java/com/flipkart/learn/cascading/cdm_data_selection/deepshit/schema/NumericFeature.java
@@ -1,14 +1,14 @@
 package com.flipkart.learn.cascading.cdm_data_selection.deepshit.schema;
 
-public class NumericFeature extends Feature<Float> {
+public class NumericFeature extends Feature<Integer> {
 
     public NumericFeature(String featureName, Source source) {
         super(featureName, source, FeatureType.numeric);
     }
 
     @Override
-    public Float clean(Object data) {
-        return ((Number)data).floatValue();
+    public Integer clean(Object data) {
+        return ((Number)data).intValue();
     }
 
 }
